Solve 2024 day 6 part 2 and expand on Grid class.
A secondary constructor allows generating individual mutable values in
the grid so it doesn't end up with the same instance in each cell.

The copy method create a shallow clone of the Grid for Grids of
immutable values.

The map method allows a deep copy or any other arbitrary transformed
grid to be created.