Solve 2022 Day 19
This was a bit of a doozy. When I finally figured out that the iteration
shouldn't go minute by minute it started to be close to working.
Instead, at each step we pick which robot to build next and then forward
through enough minutes until we can build it.

I tried to do some memoization but I don't think this problem lends
itself well to that because it depends a lot on the prior state at the
decision point. Ultimately, this got down to a reasonable runtime thanks
to some heuristics. I implemented one mentioned here:
https://old.reddit.com/r/adventofcode/comments/zpy5rm/2022_day_19_what_are_your_insights_and/

At one point I was even generating the minute-by-minute log that the
problem provided in the example. I had a bunch of annoying bugs. My loop
to move the minutes forward was not terminating correctly. The
`didBuild` var helped that. For some reason I was doing something wrong
with the base case. Still not sure what was going on there but I changed
it to handle the `> MAX_STEPS` instead of `>= MAX_STEPS` and to allow
for `max == null` and that finally gave me the right answer.