Messy solution of 2022 day 22 part 2.
Very messy because I solved it by transforming the input into a
standardized cube projection. This means that it involved a lot of
rotation transforms to the new representation and a crude transform back
into the original grid to compute the right answer.

Wrote a bunch of test code to make sure I was traversing and parsing
correctly. The `rotate` command was helpful for making expected.txt
which is my input in the standardized projection.

I think there was probably an easier way to solve this I may go back to
this at some point to try to make this less messy.