Solve 2024 day 16 part 2.
This was tricky because the Edsger couldn't be modified and using its
result to recursively recreate paths under a certain score wasn't
efficient enough.

Online recommendations suggested a DFS with a PQ to always pick the
lowest score. This works similarly to Edsger but uses the lowest score
map instead of visited.