Solve 2024 day 19 part 2
This was a matter of realizing that we can frequently reach the same
prefix in iterations via different combinations of patterns. The number
of solutions coming after a prefix is the same no matter how the prefix
was constructed because there are no constraints on how latter patterns
are picked. Therefore, we cache solves count based on prefix and use
that to prune large amounts of iterations.

The deadEnds set was abandoned because it was somehow causing us to
undercount solves. Not sure why but we can get the same data from the
cache so using a `deadEnd` helper for that part.