Improve performance of Edsger algorithm.
Previously, this was inefficient for multiple reasons:
 1. We are looking for the minimum node via a linear search.
 2. We are looking for the minimum node over all unvisited nodes (not
 just the ones that could even be reached next).
 3. We are requiring to enumerate all nodes which can explode the
 memory/time usage for graphs that are abstract (pretty much all
 graphs?) because they may consist of Point, Direction or Point, State.

To fix this all we do a few things.

We remove the need to enumerate all nodes, instead we track only the
unvisited nodes that are reachable from any that have been visited
already by progressively adding neighbors to a queue.

We use a PriorityQueue to find the minimum which is O(log n) to get the
smallest.

Also, we add a termination condition to not traverse a massive space of
possible states that are irrelevant after reaching the end.

Day16.1 needed to be updated because its solution was actually depending
on the `getNodes()` method to avoid the walls (#). Instead we just drop
it in the GetNeighbors method.