Solve 2023 Day 18 Part 2.
Jeez what a slog. I actually wrote UNIT TESTS!

It took me a while to even have an idea for what to do. I first thought:
if we try to create rectangles going inward from all boundary segments
then subtract the overlap and add the space taken up by the boundary.

This was starting to be a slog to try to compute rects (and
intersections) coming in from all directions. Then I realized that we
could just pick a direction and hardcode that. This would mean that we
only take vertical segments going EAST toward the inside of the areas
(but any direction would work). Then we don't have to worry about
overlapping rects or generalizing our intersection logic. We could just
worry about growing in the X axis and intersecting with other segments
on the X axis.

Unfortunately, there were still a ton of edge cases so like interference
(a neighboring horizontal segment) that meant we would have to shorten
the height of the rect we're building. Imagine the leftmost vertical
segment here:

 ####
 #..#
 #..#
 ####

The segment goes from (0,0) to (0, 3). The rect that would grow out of
this would be height 2 (i.e. we wouldn't want to count the horizontal
lines above or below).

But consider segment (3, 0) to (3, 3) in this case:

 #######
 #.....#
 ####..#
 ...#..#
 ...#..#
 ...####

We would want height 3 because the top horizontal segment is going the
other direction and wouldn't overlap the rect.

Rects also have to get split whenever we hit a segment that doesn't
cover the entire vertical space from the starting segment. For example,
segment (3, 2) to (3, 7) in this case:

 ##########
 #........#
 ####.....#
 ...#...###
 ...#...#..
 ...#...#..
 ...#...###
 ####.....#
 #........#
 ##########

We would only a rect of the height 6 would get blocked by the opposite
side at segment (7, 3) -> (7, 6). So we have to create one of height 4
and 2 of height 1.

I covered all of these cases in unit tests and even checked with the
part1 input. Still I was getting the wrong answer. But I finally found
out that I had an off by 1 in my loop. I wasn't closing the gap between
the end of the list and the beginning of the list looping again. I had
to actually render the output in an image (since it was way to big to
render on the terminal) so I could find the rect I was missing.